Add no-op "stage" taskt to build.gradle
This task is executed by default by Heroku for Gradle apps during the
build. However, we don't need to do anything there, because our command
for starting the app that we have in the Procfile is ./gradlew run,
which compiles and builds everything. That's why the stage task is no-op.