Modify "stage" task to build the app
In this way, the download of dependencies and compilation is done during
the build of the Heroku app, and not during the startup.

This means that the Heroku build is slower, but the app startup is
faster. This is good for starting an instance of the RPC client on a
one-off dyno with:

heroku run "./gradlew run"

In essence, the difference to before is that the application slug
produced by the Heroku build already contains the build/ folder with all
the compiled binaries, whereas in the former case, the slug did not
contain this, and it had to be created at startup time.