Use shadowJar to create a fat JAR
And then start app in Procfile with java -jar ... instead of ./gradlew run.

The reason is that there was a problem with running ./gradlew on a
one-off dyno (termcap entry not found).

Just executing a JAR file always works, and is even much faster than
starting the app with ./gradlew run.